Move definitions out of lib.rs
