+++ swap/src/network/request_response.rs
@@ -1,3 +1,4 @@
+use crate::protocol::{alice, bob};
 use async_trait::async_trait;
 use futures::prelude::*;
 use libp2p::{
@@ -7,8 +8,6 @@ use libp2p::{
 use serde::{Deserialize, Serialize};
 use std::{fmt::Debug, io, marker::PhantomData};
 use tracing::debug;
-
-use crate::protocol::{alice, bob};
 
 /// Time to wait for a response back once we send a request.
 pub const TIMEOUT: u64 = 3600; // One hour.
